src/web/requirements/base.txt
Django==4.2.2
django-filter==22.1
djangorestframework==3.14.0
django-cors-headers==3.13.0
psycopg2==2.9.5
Unipath==1.1
gunicorn==20.1.0
## The following requirements were added by pip freeze:
asgiref==3.7.2
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
asttokens==2.2.1
backcall==0.2.0
coverage==5.5
debugpy==1.6.6
decorator==5.1.1
django-extensions==3.2.1
django_debug_toolbar==3.8.1
executing==1.2.0
ipdb==0.13.13
ipython==8.14.0
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
six==1.16.0
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
wcwidth==0.2.6